Add `configure.py --warnings-as-errors` option.
Also add -DCMAKE_COMPILE_WARNING_AS_ERROR=1 to Cmake release builds.

Note that enabling them for debug builds triggers far too many warnings,
in particular in re2-generated code (e.g. lexer.cc) or in third-party one
(e.g. src/third_party/emhash/hash_table8.hpp).

See issue #2615 for context.

+ Fix a number of compiler warnings, most of them related to
  losing integer precision (e.g. from size_t to int), or ignoring
  function results.

+++ src/build.cc
@@ -951,8 +951,9 @@ bool Builder::FinishCommand(CommandRunner::Result* result, string* err) {
     disk_interface_->RemoveFile(rspfile);
 
   if (scan_.build_log()) {
-    if (!scan_.build_log()->RecordCommand(edge, start_time_millis,
-                                          end_time_millis, record_mtime)) {
+    if (!scan_.build_log()->RecordCommand(
+            edge, static_cast<int>(start_time_millis),
+            static_cast<int>(end_time_millis), record_mtime)) {
       *err = string("Error writing to build log: ") + strerror(errno);
       return false;
     }
